Extract OSCService
