+++ LivePlaylist/AbletonTrackService.swift
@@ -1,18 +1,5 @@
-import ReactiveCocoa
-
-class AbletonTrackService : NSObject, TrackService, OSCServerDelegate {
-  let client = OSCClient()
-  let server = OSCServer()
-  let (incomingMessagesSignal, incomingMessagesSink) = HotSignal<OSCMessage>.pipe()
-  
-  override init() {
-
-    super.init()
-    server.delegate = self
-    server.listen(9001)
-    
-    incomingMessagesSignal.observe { message in NSLog("All messages: \(message)") }
-  }
+class AbletonTrackService : NSObject, TrackService {
+  let osc = OSCService()
   
   func listTracks(callback: ([String]) -> ()) {
     // Pseudo-implementation:
@@ -26,23 +13,12 @@ class AbletonTrackService : NSObject, TrackService, OSCServerDelegate {
     // - receive series of /live/name/scene (int scene, string name)
     
     let message = OSCMessage(address: "/live/scenes", arguments: [])
-    client.sendMessage(message!, to: "udp://localhost:9000")
+    osc.sendMessage(message)
     
-    incomingMessagesSignal
+    osc.incomingMessagesSignal
       .take(1)
       .map { [$0.address] }
       .observe(callback)
   }
-  
-  func handleMessage(incomingMessage: OSCMessage!) {
-    if let message = incomingMessage {
-      incomingMessagesSink.put(message)
-    }
-  }
 }
 
-extension OSCMessage: Printable {
-  public override var description: String {
-    return "\(self.address) \(self.arguments)"
-  }
-}
